Full functional and tested Java wrapper for Qt 5.14

+++ generator/parser/ast.cpp
@@ -49,7 +49,7 @@
 // kate: space-indent on; indent-width 2; replace-tabs on;
 
 QString AST::toString(TokenStream *stream) const {
-    const Token &tk = stream->token((int) start_token);
-    const Token &end_tk = stream->token((int) end_token);
-    return QString::fromLatin1(tk.text + (int) tk.position, (int)(end_tk.position - tk.position));
+    const Token &tk = stream->token(start_token);
+    const Token &end_tk = stream->token(end_token);
+    return QString::fromLatin1(tk.text + int(tk.position), int(end_tk.position - tk.position));
 }
